zmiana na float

+++ PCA_CULA/pca.cuh
@@ -2,7 +2,7 @@
 #ifndef PCA_H
 #define PCA_H
 
-typedef double nifti_data_type;
+typedef float nifti_data_type;
 
 void runPCA(nifti_data_type* A, int m, int n);
